+++ src/client/views/load_queue/index.jsx
@@ -1,7 +1,6 @@
 /* Copyright G. Hemingway, 2015 - All rights reserved */
 'use strict';
 
-import React from 'react';
 require('../../stylesheets/components/load_queue.scss');
 
 /*************************************************************************/