+++ src/client/views/sidebar/workplanlist/index.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {Treebeard} from 'react-treebeard';
 import ts from '../tree_style.jsx';